function_app/requirements.txt
azure-functions>=1.17.0
azure-identity>=1.15.0
azure-cosmos>=4.5.0
azure-keyvault-secrets>=4.7.0
azure-monitor-opentelemetry>=1.2.0
aiohttp>=3.9.0
httpx>=0.27.0
PyJWT>=2.8.0
cryptography>=42.0.0
pydantic>=2.5.0
